Cast types for mrtg_scrape sprintf

+++ opentracker.c
@@ -271,9 +271,9 @@ e500:
     { 
       time_t seconds_elapsed = time( NULL ) - ot_start_time;
       reply_size = sprintf( static_reply + SUCCESS_HTTP_HEADER_LENGTH, 
-                            "%d\n%d\nUp: %ld seconds (%ld hours)\nPretuned by german engineers, currently handling %li connections per second.",
-                            ot_overall_connections, ot_overall_connections, seconds_elapsed,
-                            seconds_elapsed / 3600, ot_overall_connections / ( seconds_elapsed ? seconds_elapsed : 1 ) );
+                            "%i\n%i\nUp: %i seconds (%i hours)\nPretuned by german engineers, currently handling %i connections per second.",
+                            ot_overall_connections, ot_overall_connections, (int)seconds_elapsed,
+                            (int)(seconds_elapsed / 3600), ot_overall_connections / ( seconds_elapsed ? seconds_elapsed : 1 ) );
     }
     break;
   default: /* neither *scrape nor announce */
